Improve/etp stabilize and improve task block ux loading feedback and responsive layout fixes (#3934)
* feat(web): enhance task management components and introduce new dropdowns

- Refactored `TaskSecondaryInfo` to replace Listbox with DropdownMenu for improved UI/UX.
- Added new components: `ActiveTaskLabelsDropdown`, `ActiveTaskProjectDropdown`, and `MultipleStatusDropdown` for better task status management.
- Introduced hooks for task labels, priorities, sizes, and versions to streamline data handling.
- Improved task status dropdown functionality and integrated new utility functions for mapping task statuses.
- Cleaned up and optimized existing code for better maintainability and performance.

* fix(web): restore ActiveTaskStatusDropdown and adjust text ellipsis widths

- Reintroduced `ActiveTaskStatusDropdown` component for task status management.
- Updated text ellipsis widths in `ProjectDropDown` for improved UI consistency.

* refactor(web): reorganize task components and improve dropdown functionality

- Adjusted import paths for task-related components to enhance clarity and maintainability.
- Refactored dropdown components for task status, labels, and sizes to streamline user interactions.
- Improved styling consistency across various components, including adjustments to text ellipsis and layout properties.
- Cleaned up unused imports and optimized existing code for better performance and readability.

* fix(web): update background color in ProjectDropDown for improved UI consistency

* feat(web): enhance task dropdown components with loading states and improved styling

- Added loading indicators to `MultipleStatusDropdown`, `ActiveTaskStatusDropdown`, `ActiveTaskPropertiesDropdown`, and `ActiveTaskSizesDropdown` for better user feedback during updates.
- Updated task status class names to include text ellipsis and overflow handling for improved UI consistency.
- Introduced a new hook, `useTaskDropdownLoading`, to manage loading states across task dropdown components, enhancing user experience during asynchronous operations.
- Refactored `TaskLabels` component to integrate loading state management, ensuring a smoother interaction when updating task labels.

* feat(web): enhance task components with optimistic updates and improved loading indicators

- Updated `MultipleStatusDropdown` and `StatusDropdown` to use a new variant for the loading spinner, enhancing visual feedback during loading states.
- Refactored `TaskLabels` and `useActiveTaskStatus` hooks to implement optimistic updates, providing immediate UI feedback and error handling for task label and status changes.
- Integrated toast notifications for success and error states, improving user experience during task updates.

* feat(web): integrate toast notifications and improve loading states in task components

- Added success and error toast notifications in `ProjectDropDown` for task project updates, enhancing user feedback.
- Refactored `useActiveTaskStatus` to manage local loading states more effectively across task dropdown components.
- Updated loading indicators in `ActiveTaskStatusDropdown`, `ActiveTaskPropertiesDropdown`, and `ActiveTaskSizesDropdown` to improve user experience during updates.

* fix: deepscan issue
