+++ apps/web/core/components/auth/auth-user-task-input.tsx
@@ -24,41 +24,41 @@ export function AuthUserTaskInput({ className }: IClassName) {
 				showTaskNumber={true}
 				autoAssignTaskAuth={isTrackingEnabled}
 			/>
-			<div className="flex flex-row items-center gap-3 ml-2 lg:gap-4 md:justify-between lg:justify-start">
+			<div className="flex flex-row gap-3 items-center ml-2 lg:gap-4 md:justify-between lg:justify-start">
 				<div className="mb-4 xl:flex lg:mb-0">
 					<span className="pr-2 font-normal text-gray-500">{t('common.ESTIMATE')}:</span>
 					<TaskEstimate />
 				</div>
 
-				<div className="justify-end flex-1 hidden gap-2 md:flex md:items-center">
+				<div className="hidden flex-1 gap-2 justify-end md:flex md:items-center">
 					<ActiveTaskStatusDropdown
-						className="w-fit lg:max-w-[190px] dark:text-white dark:border-white"
+						className="w-fit lg:max-w-[190px] dark:text-white dark:border-gray-800"
 						disabled={!activeTeamTask}
-						taskStatusClassName="text-xs py-1.5 w-full dark:bg-[#1B1D22] dark:border dark:border-[#FFFFFF33] "
+						taskStatusClassName="text-xs py-1.5 w-full dark:bg-[#1B1D22] dark:border dark:border-[#FFFFFF33] overflow-hidden text-ellipsis whitespace-nowrap"
 					/>
 
 					<ActiveTaskPropertiesDropdown
-						className="w-fit lg:max-w-[190px] dark:text-white dark:border-white"
+						className="w-fit lg:max-w-[190px] dark:text-white dark:border-gray-800"
 						disabled={!activeTeamTask}
-						taskStatusClassName="w-full py-1.5 text-xs dark:bg-[#1B1D22] dark:border dark:border-[#FFFFFF33]"
+						taskStatusClassName="w-full py-1.5 text-xs dark:bg-[#1B1D22] dark:border dark:border-[#FFFFFF33] overflow-hidden text-ellipsis whitespace-nowrap"
 					/>
 
 					<ActiveTaskSizesDropdown
-						className="w-fit lg:max-w-[190px] dark:text-white dark:border-white"
+						className="w-fit lg:max-w-[190px] dark:text-white dark:border-gray-800"
 						disabled={!activeTeamTask}
-						taskStatusClassName="w-full py-1.5 text-xs dark:bg-[#1B1D22] dark:border dark:border-[#FFFFFF33]"
+						taskStatusClassName="w-full py-1.5 text-xs dark:bg-[#1B1D22] dark:border dark:border-[#FFFFFF33] overflow-hidden text-ellipsis whitespace-nowrap"
 					/>
 
 					<TaskLabels
 						task={activeTeamTask}
 						className="w-fit lg:max-w-[170px] text-xs"
 						forDetails={false}
-						taskStatusClassName="dark:bg-[#1B1D22] text-xs border dark:border-[#fff]"
+						taskStatusClassName="dark:bg-[#1B1D22] text-xs border dark:border-gray-700 overflow-hidden text-ellipsis whitespace-nowrap"
 					/>
 					{activeTeamTask && (
 						<ProjectDropDown
 							styles={{
-								container: 'rounded-xl grow text-xs !max-w-fit',
+								container: 'rounded-xl grow text-xs !max-w-fit overflow-hidden',
 								listCard: 'rounded-xl'
 							}}
 							task={activeTeamTask}
@@ -67,11 +67,11 @@ export function AuthUserTaskInput({ className }: IClassName) {
 				</div>
 				{/* <div className="grid justify-items-center md:hidden">
 					<div className="flex">
-						<ActiveTaskStatusDropdown className="w-32 mr-2" disabled={!activeTeamTask} />
+						<ActiveTaskStatusDropdown className="mr-2 w-32" disabled={!activeTeamTask} />
 						<ActiveTaskPropertiesDropdown className="w-32" disabled={!activeTeamTask} />
 					</div>
 					<div className="flex mt-2">
-						<ActiveTaskSizesDropdown className="w-32 mr-2" disabled={!activeTeamTask} />
+						<ActiveTaskSizesDropdown className="mr-2 w-32" disabled={!activeTeamTask} />
 						<TaskLabels task={activeTeamTask} className="min-w-fit lg:max-w-[170px]" forDetails={false} />
 					</div>
 				</div> */}
