+++ apps/web/core/components/tasks/task-issue.tsx
@@ -2,15 +2,6 @@ import { useModal } from '@/core/hooks';
 import { clsxm } from '@/core/lib/utils';
 import { BackButton, Button, Modal, Text } from '@/core/components';
 import { NoteIcon, BugIcon, Square4StackIcon, Square4OutlineIcon } from 'assets/svg';
-import {
-	IActiveTaskStatuses,
-	StatusDropdown,
-	TaskStatus,
-	TStatus,
-	TStatusItem,
-	TTaskVersionsDropdown,
-	useActiveTaskStatus
-} from './task-status';
 import { useTranslations } from 'next-intl';
 import { useCallback, useMemo, useState } from 'react';
 import { cn } from '@/core/lib/helpers';
@@ -24,6 +15,14 @@ import { Select } from '../features/projects/add-or-edit-project/steps/basic-inf
 import { useAtomValue } from 'jotai';
 import { issueTypesListState } from '@/core/stores';
 import Image from 'next/image';
+import {
+	IActiveTaskStatuses,
+	TStatus,
+	TStatusItem,
+	TTaskVersionsDropdown
+} from '@/core/types/interfaces/task/task-card';
+import { useActiveTaskStatus } from '@/core/hooks/tasks/use-active-task-status';
+import { StatusDropdown, TaskStatus } from './task-status';
 
 const defaultTaskClasses = 'w-full min-w-[10px] flex-none aspect-square max-w-[12px] text-white';
 export const taskIssues: TStatus<EIssueType> = {
@@ -84,7 +83,7 @@ export function TaskIssuesDropdown({
 		(item: (typeof taskIssuesOptions)[number]) => (
 			<div
 				style={{ backgroundColor: item.color ?? undefined }}
-				className="flex w-full items-center gap-2 py-1 px-2 rounded-md"
+				className="flex gap-2 items-center px-2 py-1 w-full rounded-md"
 			>
 				<div className="w-[1rem] flex items-center justify-center h-[1rem] p-[.02rem] rounded">
 					{item.fullIconUrl && (
@@ -106,7 +105,7 @@ export function TaskIssuesDropdown({
 	const renderValue = useCallback((value: string | null) => {
 		const item = taskIssuesOptions.find((el) => el.name == value);
 		return value ? (
-			<div className="flex items-center gap-2">
+			<div className="flex gap-2 items-center">
 				<div
 					style={{
 						backgroundColor: item?.color ?? undefined
@@ -128,7 +127,7 @@ export function TaskIssuesDropdown({
 			</div>
 		) : (
 			<div className="w-[1.5rem] border flex items-center justify-center h-[1.5rem] p-[.3rem] rounded-lg">
-				<div className="w-full border border-black/40 h-full rounded-full"></div>
+				<div className="w-full h-full rounded-full border border-black/40"></div>
 			</div>
 		);
 	}, []);
@@ -259,12 +258,12 @@ export function CreateTaskIssueModal({ open, closeModal }: { open: boolean; clos
 		<Modal isOpen={open} closeModal={closeModal}>
 			<form className="w-[98%] md:w-[430px]" autoComplete="off" onSubmit={handleSubmit}>
 				<EverCard className="w-full" shadow="custom">
-					<div className="flex flex-col items-center justify-between">
+					<div className="flex flex-col justify-between items-center">
 						<Text.Heading as="h3" className="text-center">
 							{t('common.CREATE_ISSUE')}
 						</Text.Heading>
 
-						<div className="w-full mt-5">
+						<div className="mt-5 w-full">
 							<InputField
 								name="name"
 								autoCustomFocus
@@ -273,7 +272,7 @@ export function CreateTaskIssueModal({ open, closeModal }: { open: boolean; clos
 							/>
 						</div>
 
-						<div className="flex items-center justify-between w-full mt-3">
+						<div className="flex justify-between items-center mt-3 w-full">
 							<BackButton onClick={closeModal} />
 							<Button type="submit">{t('common.CREATE')}</Button>
 						</div>
