+++ apps/web/core/components/tasks/task-labels.tsx
@@ -2,14 +2,16 @@
 
 import { useModal, useSyncRef, useTaskLabels, useTeamTasks } from '@/core/hooks';
 import { Button, Modal } from '@/core/components';
-import { TaskLabelsDropdown, taskUpdateQueue } from './task-status';
+import { TaskLabelsDropdown } from '@/core/components/tasks/task-status';
 import { debounce, isEqual } from 'lodash';
-import { useCallback, useMemo, useRef } from 'react';
+import { useCallback, useMemo, useRef, useState, useEffect } from 'react';
+import { toast } from 'sonner';
 import { AddIcon } from 'assets/svg';
 import { TaskLabelForm } from './task-labels-form';
 import { EverCard } from '../common/ever-card';
 import { Nullable } from '@/core/types/generics/utils';
 import { TTask } from '@/core/types/schemas/task/task.schema';
+import { taskUpdateQueue } from '@/core/utils/task.utils';
 
 type Props = {
 	task: Nullable<TTask>;
@@ -24,6 +26,13 @@ export function TaskLabels({ task, className, forDetails, taskStatusClassName, o
 	const { taskLabels } = useTaskLabels();
 	const modal = useModal();
 	const latestLabels = useRef<string[]>([]);
+	const [isLoading, setIsLoading] = useState(false);
+	const [optimisticLabels, setOptimisticLabels] = useState<string[] | null>(null);
+
+	// Reset optimistic labels when task changes
+	useEffect(() => {
+		setOptimisticLabels(null);
+	}, [task?.id]);
 
 	const onValuesChange = useCallback(
 		(_: any, values: string[] | undefined) => {
@@ -33,13 +42,32 @@ export function TaskLabels({ task, className, forDetails, taskStatusClassName, o
 				return;
 			}
 
+			// Set optimistic value immediately
+			setOptimisticLabels(values || []);
+			setIsLoading(true);
+
 			taskUpdateQueue.task(
 				(task, taskLabels, values) => {
 					return updateTask({
 						// eslint-disable-next-line  @typescript-eslint/no-non-null-assertion
 						...task.current!,
 						tags: taskLabels.filter((tag) => (tag.name ? values?.includes(tag.name) : false)) as any
-					});
+					})
+						.then(() => {
+							// Success - keep optimistic value and show success toast
+							toast.success('Task labels updated successfully');
+						})
+						.catch((error) => {
+							// Error - revert optimistic value and show error toast
+							setOptimisticLabels(null);
+							toast.error('Failed to update task labels', {
+								description: error?.message || 'Please try again'
+							});
+							console.error('Error updating task labels:', error);
+						})
+						.finally(() => {
+							setIsLoading(false);
+						});
 				},
 				$task,
 				taskLabels,
@@ -51,7 +79,9 @@ export function TaskLabels({ task, className, forDetails, taskStatusClassName, o
 
 	const onValuesChangeDebounce = useMemo(() => debounce(onValuesChange, 2000), [onValuesChange]);
 
-	const tags = (task?.tags as typeof taskLabels | undefined)?.map((tag) => tag.name || '');
+	// Use optimistic labels if available, otherwise use actual task labels
+	const actualTags = (task?.tags as typeof taskLabels | undefined)?.map((tag) => tag.name || '');
+	const tags = optimisticLabels !== null ? optimisticLabels : actualTags;
 
 	return (
 		<>
@@ -65,9 +95,10 @@ export function TaskLabels({ task, className, forDetails, taskStatusClassName, o
 				sidebarUI={forDetails}
 				taskStatusClassName={taskStatusClassName}
 				latestLabels={latestLabels}
+				isLoading={isLoading}
 			>
 				<Button
-					className="w-full px-2 py-1 mt-4 text-xs dark:text-white dark:border-white"
+					className="px-2 py-1 mt-4 w-full text-xs dark:text-white dark:border-white"
 					variant="outline"
 					onClick={modal.openModal}
 				>
