+++ apps/web/core/hooks/tasks/use-task-priorities-value.ts
@@ -1,0 +1,8 @@
+import { TTaskStatus } from '@/core/types/schemas';
+import { useMapToTaskStatusValues } from './use-map-to-task-status-values';
+import { useTaskPriorities } from './use-task-priorities';
+
+export function useTaskPrioritiesValue() {
+	const { taskPriorities } = useTaskPriorities();
+	return useMapToTaskStatusValues(taskPriorities as TTaskStatus[], false);
+}
