+++ apps/web/core/utils/task.utils.ts
@@ -1,0 +1,3 @@
+import { Queue } from '../lib/utils';
+
+export const taskUpdateQueue = new Queue(1);
